requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
click==8.1.8
colorama==0.4.6
databases==0.9.0
distlib==0.3.9
Django==5.1.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1

dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
filelock==3.16.1
fpdf==1.7.2
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
mysql-connector-python==9.1.0
mysqlclient==2.2.6
packaging==25.0
platformdirs==4.3.6
pydantic==2.10.5
pydantic_core==2.27.2
PyMySQL==1.1.1
pyodbc==5.2.0
python-decouple==3.8
pytz==2024.2
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlmodel==0.0.22
sqlparse==0.5.3
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
virtualenv==20.28.0
whitenoise==6.9.0
requests   2.31.0
